# swagger/swagger.yml
openapi: 3.0.0
info:
  title: "AllTrails Users RESTful API"
  description: "ProfileService Microservice API for AllTrails"
  version: 1.0.0

# add security
security:
  - basicAuth: []
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: authenticator.authenticate_user
...
paths:
...
  /api/admin/accounts:
    get:
      security:
        - basicAuth: []
      operationId: python_endpoints.administrator.getAllUsers
      tags:
        - Admin
      summary: "Show all user accounts"
      responses:
        "200":
          description: "Successfully retrieved all user accounts"
        "401":
          description: "Unauthorised access"        